chore(utils): Move cf module (#1831)
* move cloudflare

* typo

* fix import order

* rename

* missing import
